+++ src/contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper/3.0.0.ts
@@ -1,140 +1,11 @@
-import {
-  CountTransferManagerContract_3_0_0,
-  CountTransferManagerEventArgs_3_0_0,
-  CountTransferManagerEvents_3_0_0,
-  CountTransferManagerModifyHolderCountEventArgs_3_0_0,
-  CountTransferManagerPauseEventArgs_3_0_0,
-  CountTransferManagerUnpauseEventArgs_3_0_0,
-  LogWithDecodedArgs,
-  Web3Wrapper,
-} from '@polymathnetwork/abi-wrappers';
-import { schemas } from '@0x/json-schemas';
-import assert from '../../../../utils/assert';
-import {
-  ContractVersion,
-  SubscribeAsyncParams,
-  GetLogsAsyncParams,
-  Subscribe,
-  GetLogs,
-  EventCallback,
-} from '../../../../types';
-import ContractFactory from '../../../../factories/contractFactory';
+import { ContractVersion } from '../../../../types';
 
 import CountTransferManagerCommon from './common';
 
-interface ModifyHolderCountSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.ModifyHolderCount;
-  callback: EventCallback<CountTransferManagerModifyHolderCountEventArgs_3_0_0>;
-}
-
-interface GetModifyHolderCountLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.ModifyHolderCount;
-}
-
-interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Pause;
-  callback: EventCallback<CountTransferManagerPauseEventArgs_3_0_0>;
-}
-
-interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Pause;
-}
-
-interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Unpause;
-  callback: EventCallback<CountTransferManagerUnpauseEventArgs_3_0_0>;
-}
-
-interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
-  eventName: CountTransferManagerEvents_3_0_0.Unpause;
-}
-
-interface CountTransferManagerSubscribeAsyncParams extends Subscribe {
-  (params: ModifyHolderCountSubscribeAsyncParams): Promise<string>;
-  (params: PauseSubscribeAsyncParams): Promise<string>;
-  (params: UnpauseSubscribeAsyncParams): Promise<string>;
-}
-
-interface GetCountTransferManagerLogsAsyncParams extends GetLogs {
-  (params: GetModifyHolderCountLogsAsyncParams): Promise<
-    LogWithDecodedArgs<CountTransferManagerModifyHolderCountEventArgs_3_0_0>[]
-  >;
-  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<CountTransferManagerPauseEventArgs_3_0_0>[]>;
-  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<CountTransferManagerUnpauseEventArgs_3_0_0>[]>;
-}
-
-export class CountTransferManager_3_0_0 extends CountTransferManagerCommon {
-  public contract: Promise<CountTransferManagerContract_3_0_0>;
-
-  public contractVersion = ContractVersion.V3_0_0;
-
-  /**
-   * Instantiate CountTransferManager_3_0_0
-   * @param web3Wrapper Web3Wrapper instance to use
-   * @param contract
-   * @param contractFactory
-   */
-  public constructor(
-    web3Wrapper: Web3Wrapper,
-    contract: Promise<CountTransferManagerContract_3_0_0>,
-    contractFactory: ContractFactory,
-  ) {
-    super(web3Wrapper, contract, contractFactory);
-    this.contract = contract;
-  }
-
-  /**
-   * Subscribe to an event type emitted by the contract.
-   * @return Subscription token used later to unsubscribe
-   */
-  public subscribeAsync: CountTransferManagerSubscribeAsyncParams = async <
-    ArgsType extends CountTransferManagerEventArgs_3_0_0
-  >(
-    params: SubscribeAsyncParams,
-  ): Promise<string> => {
-    assert.doesBelongToStringEnum('eventName', params.eventName, CountTransferManagerEvents_3_0_0);
-    assert.doesConformToSchema('indexFilterValues', params.indexFilterValues, schemas.indexFilterValuesSchema);
-    assert.isFunction('callback', params.callback);
-    const normalizedContractAddress = (await this.contract).address.toLowerCase();
-    const subscriptionToken = await this.subscribeInternal<ArgsType>(
-      normalizedContractAddress,
-      params.eventName,
-      params.indexFilterValues,
-      params.callback,
-      params.isVerbose,
-    );
-    return subscriptionToken;
-  };
-
-  /**
-   * Gets historical logs without creating a subscription
-   * @return Array of logs that match the parameters
-   */
-  public getLogsAsync: GetCountTransferManagerLogsAsyncParams = async <
-    ArgsType extends CountTransferManagerEventArgs_3_0_0
-  >(
-    params: GetLogsAsyncParams,
-  ): Promise<LogWithDecodedArgs<ArgsType>[]> => {
-    assert.doesBelongToStringEnum('eventName', params.eventName, CountTransferManagerEvents_3_0_0);
-    const normalizedContractAddress = (await this.contract).address.toLowerCase();
-    const logs = await this.getLogsAsyncInternal<ArgsType>(
-      normalizedContractAddress,
-      params.eventName,
-      params.blockRange,
-      params.indexFilterValues,
-    );
-    return logs;
-  };
-}
+export class CountTransferManager_3_0_0 extends CountTransferManagerCommon {}
 
 export function isCountTransferManager_3_0_0(
   wrapper: CountTransferManagerCommon,
 ): wrapper is CountTransferManager_3_0_0 {
   return wrapper.contractVersion === ContractVersion.V3_0_0;
-}
-
-export function isCountTransferManager_3_1_0(
-  wrapper: CountTransferManagerCommon,
-): wrapper is CountTransferManager_3_0_0 {
-  return wrapper.contractVersion === ContractVersion.V3_1_0;
 }
