+++ src/contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper/__tests__/3.0.0.test.ts
@@ -1,14 +1,13 @@
 // CountTransferManager test
-import { mock, instance, reset, when, verify, objectContaining } from 'ts-mockito';
+import { mock, instance, reset } from 'ts-mockito';
 import {
   CountTransferManagerContract_3_0_0,
   ISecurityTokenContract_3_0_0,
-  PolyTokenEvents_3_0_0,
-  BigNumber,
   Web3Wrapper,
 } from '@polymathnetwork/abi-wrappers';
 import ContractFactory from '../../../../../factories/contractFactory';
 import { CountTransferManager_3_0_0 } from '../3.0.0';
+import CountTransferManagerCommon from '../common';
 
 describe('CountTransferManagerWrapper', () => {
   let target: CountTransferManager_3_0_0;
@@ -38,20 +37,9 @@ describe('CountTransferManagerWrapper', () => {
     reset(mockedSecurityTokenContract);
   });
 
-  describe('SubscribeAsync', () => {
-    test('should throw as eventName does not belong to CountTransferManager', async () => {
-      // Mocked parameters
-      const mockedParams = {
-        eventName: PolyTokenEvents_3_0_0.Transfer,
-        indexFilterValues: {},
-        callback: () => {},
-        isVerbose: false,
-      };
-
-      // Real call
-      await expect(target.subscribeAsync(mockedParams)).rejects.toEqual(
-        new Error(`Expected eventName to be one of: 'ModifyHolderCount', 'Pause', 'Unpause', encountered: Transfer`),
-      );
+  describe('Types', () => {
+    test('should extend CountTransferManagerCommon', async () => {
+      expect(target instanceof CountTransferManagerCommon).toBe(true);
     });
   });
 });
