+++ src/contract_wrappers/modules/transfer_manager/count_transfer_manager_wrapper/common.ts
@@ -1,14 +1,72 @@
 import {
   CountTransferManagerContract_3_0_0,
+  CountTransferManagerEventArgs_3_0_0,
+  CountTransferManagerEvents_3_0_0,
+  CountTransferManagerModifyHolderCountEventArgs_3_0_0,
+  CountTransferManagerPauseEventArgs_3_0_0,
+  CountTransferManagerUnpauseEventArgs_3_0_0,
   Web3Wrapper,
   BigNumber,
   PolyResponse,
+  LogWithDecodedArgs,
 } from '@polymathnetwork/abi-wrappers';
+import { schemas } from '@0x/json-schemas';
+import assert from '../../../../utils/assert';
 import { numberToBigNumber, parseTransferResult, valueToWei } from '../../../../utils/convert';
 import ContractFactory from '../../../../factories/contractFactory';
-import { ErrorCode, Perm, TransferResult, TxParams } from '../../../../types';
+import {
+  ErrorCode,
+  Perm,
+  TransferResult,
+  TxParams,
+  SubscribeAsyncParams,
+  GetLogsAsyncParams,
+  Subscribe,
+  GetLogs,
+  EventCallback,
+} from '../../../../types';
 import ModuleWrapper from '../../module_wrapper';
-import assert from '../../../../utils/assert';
+
+interface ModifyHolderCountSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: CountTransferManagerEvents_3_0_0.ModifyHolderCount;
+  callback: EventCallback<CountTransferManagerModifyHolderCountEventArgs_3_0_0>;
+}
+
+interface GetModifyHolderCountLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: CountTransferManagerEvents_3_0_0.ModifyHolderCount;
+}
+
+interface PauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: CountTransferManagerEvents_3_0_0.Pause;
+  callback: EventCallback<CountTransferManagerPauseEventArgs_3_0_0>;
+}
+
+interface GetPauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: CountTransferManagerEvents_3_0_0.Pause;
+}
+
+interface UnpauseSubscribeAsyncParams extends SubscribeAsyncParams {
+  eventName: CountTransferManagerEvents_3_0_0.Unpause;
+  callback: EventCallback<CountTransferManagerUnpauseEventArgs_3_0_0>;
+}
+
+interface GetUnpauseLogsAsyncParams extends GetLogsAsyncParams {
+  eventName: CountTransferManagerEvents_3_0_0.Unpause;
+}
+
+interface CountTransferManagerSubscribeAsyncParams extends Subscribe {
+  (params: ModifyHolderCountSubscribeAsyncParams): Promise<string>;
+  (params: PauseSubscribeAsyncParams): Promise<string>;
+  (params: UnpauseSubscribeAsyncParams): Promise<string>;
+}
+
+interface GetCountTransferManagerLogsAsyncParams extends GetLogs {
+  (params: GetModifyHolderCountLogsAsyncParams): Promise<
+    LogWithDecodedArgs<CountTransferManagerModifyHolderCountEventArgs_3_0_0>[]
+  >;
+  (params: GetPauseLogsAsyncParams): Promise<LogWithDecodedArgs<CountTransferManagerPauseEventArgs_3_0_0>[]>;
+  (params: GetUnpauseLogsAsyncParams): Promise<LogWithDecodedArgs<CountTransferManagerUnpauseEventArgs_3_0_0>[]>;
+}
 
 export namespace CountTransferManagerTransactionParams {
   export interface ChangeHolderCount extends ChangeHolderCountParams {}
@@ -139,5 +197,48 @@ export default class CountTransferManagerCommon extends ModuleWrapper {
       params.txData,
       params.safetyFactor,
     );
+  };
+
+  /**
+   * Subscribe to an event type emitted by the contract.
+   * @return Subscription token used later to unsubscribe
+   */
+  public subscribeAsync: CountTransferManagerSubscribeAsyncParams = async <
+    ArgsType extends CountTransferManagerEventArgs_3_0_0
+  >(
+    params: SubscribeAsyncParams,
+  ): Promise<string> => {
+    assert.doesBelongToStringEnum('eventName', params.eventName, CountTransferManagerEvents_3_0_0);
+    assert.doesConformToSchema('indexFilterValues', params.indexFilterValues, schemas.indexFilterValuesSchema);
+    assert.isFunction('callback', params.callback);
+    const normalizedContractAddress = (await this.contract).address.toLowerCase();
+    const subscriptionToken = await this.subscribeInternal<ArgsType>(
+      normalizedContractAddress,
+      params.eventName,
+      params.indexFilterValues,
+      params.callback,
+      params.isVerbose,
+    );
+    return subscriptionToken;
+  };
+
+  /**
+   * Gets historical logs without creating a subscription
+   * @return Array of logs that match the parameters
+   */
+  public getLogsAsync: GetCountTransferManagerLogsAsyncParams = async <
+    ArgsType extends CountTransferManagerEventArgs_3_0_0
+  >(
+    params: GetLogsAsyncParams,
+  ): Promise<LogWithDecodedArgs<ArgsType>[]> => {
+    assert.doesBelongToStringEnum('eventName', params.eventName, CountTransferManagerEvents_3_0_0);
+    const normalizedContractAddress = (await this.contract).address.toLowerCase();
+    const logs = await this.getLogsAsyncInternal<ArgsType>(
+      normalizedContractAddress,
+      params.eventName,
+      params.blockRange,
+      params.indexFilterValues,
+    );
+    return logs;
   };
 }
